system_daemon/requirements.txt
grpcio
protobuf
util-base
psutil
uptime
py-dmidecode
python-daemon
